rename files of reducers and actions
