+++ src/actions/changeStatus.action.js
@@ -1,4 +1,4 @@
-import { CHANGE_STATUS } from '../../constants/constants';
+import { CHANGE_STATUS } from '../constants/constants';
 
 export const changeStatus = (id) => ({
   type: CHANGE_STATUS,
