+++ src/reducers/changeStatus.reducer.js
@@ -1,5 +1,5 @@
-import { CHANGE_STATUS } from '../../constants/constants';
-import jsonData from '../../data/data.json';
+import { CHANGE_STATUS } from '../constants/constants';
+import jsonData from '../data/data.json';
 
 const changeStatus = (state = jsonData, action) => {
   switch (action.type) {
